feat: 기한에 따른 result 페이지 조건부 렌더링
