+++ src/views/ResultPage/components/style.css.ts
